requirements.txt
pytest
discord.py
